+++ NEWS.md
@@ -3,6 +3,7 @@
 * Improvements in estimation of forecast accuracy with rolling origin evaluation.
 * The way in which preprocessings are specified has changed.
 * Method plot.utsf is implemented.
+* Linear models (stats::lm) are supported.
 
 # utsf 1.0.0